Remove conflicting interop structs
Removed our `RECT`, `HDC`, etc structs in favor of cswin32. This necessitated changing all of the existing P/Invokes back to `[DllImport]` as the `[LibaryImport]` doesn't work with the source generated types.

- Remove all `using foo =` workarounds.
- Push forward to cswin32 daily build to address bool cast issue
- New cswin32 uses System.Drawing types where plausible (`POINT` -> `Point`, as we currently do).

+++ src/System.Windows.Forms.Design/src/System/ComponentModel/Design/DesignerActionUI.cs
@@ -1091,7 +1091,7 @@ namespace System.ComponentModel.Design
 
         private void WmActivate(ref Message m)
         {
-            if ((User32.WA)m.WParamInternal == User32.WA.INACTIVE)
+            if ((User32.WA)(nint)m.WParamInternal == User32.WA.INACTIVE)
             {
                 IntPtr hwndActivating = m.LParamInternal;
                 if (WindowOwnsWindow(Handle, hwndActivating))
